Move admin export to its own concern for #3644
